backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic[email]==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.0
boto3==1.28.50
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
websockets==11.0.3
langchain==0.0.300
openai==1.3.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
asyncpg
